added ui functionality and altered classes for efficiency

+++ HDD.java
@@ -1,13 +1,13 @@
 public class HDD {
-    private float hddSize;
+    private int hddSize;
     private String hddSizeUnit;
 
-    HDD(float hddSize) {
+    HDD(int hddSize) {
         this.hddSize = hddSize;
         this.hddSizeUnit = "GB";
     }
     
     public String printHddDetails() {
-        return Float.toString(hddSize) + " " + hddSizeUnit;
+        return Integer.toString(hddSize) + " " + hddSizeUnit;
     }
 }
